Svasta nesto,citaj notion 4

+++ app/src/main/java/rs/raf/vezbe11/data/models/responses/CategoriesResponse.kt
@@ -5,7 +5,7 @@ import com.squareup.moshi.JsonClass
 @JsonClass(generateAdapter = true)
 class CategoriesResponse (
     //imena su bas kao i u apiju
-    val idCategory:Int,
+    val idCategory:Long,
     val strCategory:String,
     val strCategoryThumb:String,
     val strCategoryDescription:String,
